Govind Assignment/Question 2/creation/custom_emp/requirements.txt
Django==3.1.3
djangorestframework==3.12.1

